updater/requirements.txt
python-rtmidi==1.5.0
spsdk==1.8.0
pyusb
